extract old screen and create new screen
