+++ src/lib/Scenes/MyCollection/Screens/Artwork/MyCollectionArtwork.tests.tsx
@@ -1,92 +1,47 @@
-import { ActionType, ContextModule, editCollectedArtwork, OwnerType } from "@artsy/cohesion"
-import { FancyModalHeader } from "lib/Components/FancyModal/FancyModalHeader"
-import { navigate } from "lib/navigation/navigate"
-import { GlobalStore } from "lib/store/GlobalStore"
-import { mockTrackEvent } from "lib/tests/globallyMockedStuff"
-import { renderWithWrappers } from "lib/tests/renderWithWrappers"
+import { __globalStoreTestUtils__ } from "lib/store/GlobalStore"
+import { renderWithWrappersTL } from "lib/tests/renderWithWrappers"
 import React from "react"
-import { MyCollectionArtworkInsightsFragmentContainer } from "./Components/ArtworkInsights/MyCollectionArtworkInsights"
-import { MyCollectionArtworkHeaderFragmentContainer } from "./Components/MyCollectionArtworkHeader"
-import { MyCollectionArtworkMetaFragmentContainer } from "./Components/MyCollectionArtworkMeta"
-import { tests } from "./MyCollectionArtwork"
+import { MyCollectionArtworkQueryRenderer } from "./MyCollectionArtwork"
 
-jest.mock("./Components/MyCollectionArtworkHeader", () => ({
-  MyCollectionArtworkHeaderFragmentContainer: () => null,
-}))
-
-jest.mock("./Components/MyCollectionArtworkMeta", () => ({
-  MyCollectionArtworkMetaFragmentContainer: () => null,
-}))
-
-jest.mock("./Components/ArtworkInsights/MyCollectionArtworkInsights", () => ({
-  MyCollectionArtworkInsightsFragmentContainer: () => null,
-}))
-
-describe("MyCollectionArtworkDetail", () => {
-  const getWrapper = (props?: any) => {
-    return renderWithWrappers(<tests.MyCollectionArtwork {...props} />)
+jest.mock("./OldMyCollectionArtwork.tsx", () => {
+  const View = require("react-native/Libraries/Components/View/View")
+  return {
+    OldMyCollectionArtworkQueryRenderer: () => <View testID="old-my-collection-artwork" />,
   }
+})
 
-  afterEach(() => {
-    jest.clearAllMocks()
+describe("My Collection Artwork", () => {
+  it("show old my collection artwork page when the feature flag is disabled", () => {
+    __globalStoreTestUtils__?.injectFeatureFlags({ AREnableNewMyCollectionArtwork: false })
+
+    const { getByTestId } = renderWithWrappersTL(
+      <MyCollectionArtworkQueryRenderer
+        artworkSlug="random-slug"
+        artistInternalID="internal-id"
+        medium="medium"
+      />
+    )
+
+    expect(getByTestId("old-my-collection-artwork")).toBeTruthy()
+    expect(() => getByTestId("my-collection-artwork")).toThrowError(
+      "Unable to find an element with testID: my-collection-artwork"
+    )
   })
 
-  describe("artwork detail behavior", () => {
-    it("renders correct components", () => {
-      const artworkProps = { artwork: { internalID: "someInternalId" } }
-      const wrapper = getWrapper(artworkProps)
-      expect(wrapper.root.findByType(MyCollectionArtworkHeaderFragmentContainer)).toBeDefined()
-      expect(wrapper.root.findByType(MyCollectionArtworkMetaFragmentContainer)).toBeDefined()
-      expect(wrapper.root.findByType(MyCollectionArtworkInsightsFragmentContainer)).toBeDefined()
-    })
+  it("show new my collection artwork page when the feature flag is disabled", () => {
+    __globalStoreTestUtils__?.injectFeatureFlags({ AREnableNewMyCollectionArtwork: true })
 
-    it("calls startEditingArtworkAction when header edit button is pressed", () => {
-      const artworkProps = { artwork: { internalID: "someInternalId" } }
-      const spy = jest.fn()
-      GlobalStore.actions.myCollection.artwork.startEditingArtwork = spy as any
-      const wrapper = getWrapper(artworkProps)
-      wrapper.root.findByType(FancyModalHeader).props.onRightButtonPress()
-      expect(spy).toHaveBeenCalledWith(artworkProps.artwork)
-    })
+    const { getByTestId } = renderWithWrappersTL(
+      <MyCollectionArtworkQueryRenderer
+        artworkSlug="random-slug"
+        artistInternalID="internal-id"
+        medium="medium"
+      />
+    )
 
-    it("navigates to sales page when learn more button is pressed", () => {
-      const artworkProps = { artwork: { internalID: "someInternalId" } }
-      const wrapper = getWrapper(artworkProps)
-      wrapper.root.findByProps({ testID: "LearnMoreButton" }).props.onPress()
-      expect(navigate).toHaveBeenCalledWith("/sales")
-    })
-
-    // Analytics
-
-    it("tracks an analytics event when edit button is pressed", () => {
-      const artworkProps = { artwork: { internalID: "someInternalId", slug: "someSlug" } }
-      GlobalStore.actions.myCollection.artwork.startEditingArtwork = jest.fn() as any
-
-      const wrapper = getWrapper(artworkProps)
-      wrapper.root.findByType(FancyModalHeader).props.onRightButtonPress()
-
-      expect(mockTrackEvent).toHaveBeenCalledTimes(1)
-      expect(mockTrackEvent).toHaveBeenCalledWith(
-        editCollectedArtwork({ contextOwnerId: "someInternalId", contextOwnerSlug: "someSlug" })
-      )
-    })
-
-    it("tracks an analytics event learn more button is pressed", () => {
-      const artworkProps = { artwork: { internalID: "someInternalId", slug: "someSlug" } }
-      GlobalStore.actions.myCollection.artwork.startEditingArtwork = jest.fn() as any
-
-      const wrapper = getWrapper(artworkProps)
-      wrapper.root.findByProps({ testID: "LearnMoreButton" }).props.onPress()
-
-      expect(mockTrackEvent).toHaveBeenCalledTimes(1)
-      expect(mockTrackEvent).toHaveBeenCalledWith({
-        action: ActionType.tappedShowMore,
-        context_module: ContextModule.sellFooter,
-        context_screen_owner_type: OwnerType.myCollectionArtwork,
-        context_screen_owner_id: "someInternalId",
-        context_screen_owner_slug: "someSlug",
-        subject: "Learn More",
-      })
-    })
+    expect(getByTestId("my-collection-artwork")).toBeTruthy()
+    expect(() => getByTestId("old-my-collection-artwork")).toThrowError(
+      "Unable to find an element with testID: old-my-collection-artwork"
+    )
   })
 })
